Upgrade to listen 1.1.x.
This required a substantial rewrite of the watching code because we
needed to move to using absolute values. Additionally, there were a
number of bugs in the watcher that would cause changes in dependencies
to go unnoticed because the Sass load paths were not being watched
for changes. Additionally, the way this code was written, it was trying
to only report a change when the specific file being watched changed.
However, this code never actually worked in practice because the
subdirectory of the changed file wasn't necessarily the directory of the
watched file, and even when it was, the way it was reported from the
listener caused the subdirectory to be blank. There was also some very
bad things that would occur if the sass cache directory was a sub
directory of a watched directory.